+++ Web/ICIMS.Application/BusinessManages/Dtos/PayAuditEditDto.cs
@@ -132,7 +132,7 @@ namespace  ICIMS.BusinessManages.Dtos
 		/// <summary>
 		/// AuditUserId
 		/// </summary>
-		public int? AuditUserId { get; set; }
+		public long? AuditUserId { get; set; }
