+++ Web/ICIMS.Application/BusinessManages/Dtos/ReViewDefineEditDto.cs
@@ -66,7 +66,7 @@ namespace  ICIMS.BusinessManages.Dtos
 		/// <summary>
 		/// AuditUserId
 		/// </summary>
-		public int? AuditUserId { get; set; }
+		public long? AuditUserId { get; set; }
